Update train_sd15_lora.py
The results of training the previous set of parameters on the simpler captions was pretty bad. Didn't even look like Mario levels. This change reverts most of the previous changes.

HOWEVER: the number of iterations is still high (500 vs 100), and although I reverted the lora rank to 16, the lora alpha is now 32 (previous good results were with 16)

So, if this produces bad results, set epochs to 100 and lora_alpha to 16