Going to delete this
Tried to replace the many checkboxes with a more sophisticated interface. It is cool in some ways, but it does not fully work, and I don't like the effort of having to figure out if a given caption even exists in the training data. I'm going to revert this after committing, but wanted to keep the code in the history just in case.