ASCII data browser works
The ASCII data browser seems to work great. At first, I thought that it had revealed a problem with the data collection process, but the mistakes in the data seem to be in the original VGLC! This is a bit shocking and annoying that everyone has been using this bad data. I'm not sure it is worth fixing.