Switch to one-hot in the dataset
Clause wants to switch the level scene data to one-hot encoding in the dataset file directly, which makes sense. However, it is also switching captions to tensors, and also seems a bit confused about how I might use the level scenes if not for diffusion. I'll be making some changes to this after committing