Share code for tileset
Since tileset has to be modified with these extra bogus tiles, it became more important to avoid duplicating the code for loading the information. It's weird to return so many items though. There should probably be some kind of tileset object/class instead, expecially since the first item of the 4-tuple is apparently never needed #24