Width divisible by 4 for unconditional
Apparently, the cross attention of the text conditioned model allows the input/output size to be almost anything, but the unconditional model is much more strict with its upsampling and downsampling. For the particular architecture I use, this results in a restriction: the original input dimensions must be divisible by 4. That restriction is now captured in the genome mutation. This will restrict the conditional model's evolution too, but allows for fair comparison. Arbitrary sizes (divisible by 4) are still possible. #35 #28