Slight refactor completed
Created the prepare_conditioned_batch method, which removes some unnecessary duplication of code in the training and validation of the diffusion model. This method also incorporated the necessary changes to allow for it to use a pretrained model, though there are still some errors yet to be fixed on that front.
#42