Added for newer language models
#42
I had to add this to the requirements file in order to make the sentence transformer model successfully download on my home machine, but it's possible that my machine's environment simply had an older version of timm. I think that timm is actually a dependency of some other library already in the project. We might be able to remove this version number later.