Fixed an issue with run_diffusion with a pretraind split text encoder
When using a pretrained text encoder with split captions, there was an error when trying to generate unconditional samples. This is due to sentence_transformers_helper.py treating the split captions like a batch, and the code is unable to find the first index in an empty string. This can be fixed by simply defaulting to padding the empty string once, so it can encode properly. #68