Problem with evaluate caption adherence batch
It was hard-coded to use the SMB1 data files when it should have been matching the game data it was training on. In addition to fixing this, I made it so that the save locations resulting from re-running evaluate_caption_adherence_multi.bat will be different, so that we'll still have the old results. The results that were gathered aren't terrible to have, but simply are not the appropriate ones to calculate.