GPT metrics include astar on full level
calculate_gpt2_metrics.py now uses astar on both the smaller sampled captions and the full levels that MarioGPT generates. This should be saved in a new directory, with a name defaulting to MarioGPT_metrics. The batch file has been updated to run these metrics, as they can take a while (It's pretty much only astar). #52