FDM now saves the best model in the main directory
FDM now finally saves everything in the base model directory, and saves the model with the best caption score. There are a few things that might need to be changed to work with this new system, the only one I'm aware of is in evaluate_caption_adherance, but there might be more. For now, I'm leaving it so our normal data still works. #105