GPT long data fixed
Fixed the metric evaluation for long levels from MarioGPT in 2 spots. 1. The stride calculation was not done correctly, causing levels of a large stride and width to not be loaded. 2. average_min_edit_distance_from_real doesn't work with our longer levels, because they are a different length from our training data, so I added a case to not attempt to do this if the level is the wrong size. #52