Added a generic pipeline loader
Added a generic pipeline loader that loads the correct pipeline based off of the folders within it. If it has a unet and text encoder, it's the TextConditionalDDPMPipeline. A unet with no text encoder is the UnconditionalDDPMPipeline. If it's neither, it must be the FDMPipeline, but it does a quick check for the final-model directory, loading from there if it exists, loading from the base folder if it doesn't, to support both the legacy and current FDM saving systems. #107