Level parsing beginning to take structure
The level parsing algorithm is getting there, slowly. I'm running into an issue where almost any turn in a level causes a crash, because Mega Man levels expect the screen to cover several wall tiles on both sides. I think this can be solved with some tweaking (The best solution is to just keep moving the camera until an @ or non-solid tile is reached). However, if the game did not require this, I believe this would be working. This should be done at some point soon. #38