ascii_data_browser mostly works with megaman data
A few new updates to ascii_data_browser.py, the most notable being the addition of a new dropdown menu to select a game. This allows MegaMan data to be displayed. There's a second major change, which is in level_dataset. There's a new global variable to track the name of the tilesheet being used to allow for the code to change the used tileset halfway through execution while still not loading a new mapsheet every single time we try to generate a new level image. There is still one major bug in the ascii representation, but I'm working on it. #38 #128