Floors and ceilings captions added
Added support for floors and ceilings. Both systems work with slightly altered versions of the analyze_ceiling and analyze_floor methods from util, to avoid counting blocks such as spikes as floors or ceilings. Both of these checks only fire when exiting horizontally, so we don't mistake a platform as a ceiling with a gap. Captioning should in theory be done now, the next step is to get a model training. #38 #127