Temp bigger font/display for GUI
There was a complaint about the readability of a figure showing a picture of the GUI. I made a lot of changes to fit elements on the screen better and make the fonts bigger. I also removed some widgets taking up space. I'm going to undo this commit to revert the changes afterward, but this was what I had to do in response to the AIIDE editors.