Massive instruction updates
Trying to make code easier to use for the Evaluating Artifacts bit of AIIDE. However, I think more details are needed for the MarioGPT section on how to install/use the original MarioGPT model ... but maybe not? Maybe it just gets downloaded from Hugging Face, no problem. I'll need to try the process on a fresh machine.